Added scripts for logging jobs
These are non-interactive scripts, for logging (well, flagging-up, really...) jobs on shared machines.

They assume that a /etc/bioinfo.jobs.list file is globally read-able and write-able. Essentially, the `i-want-a-jobby` script for users to post notices of intended analyses to the list, while the `profile` and `zlogin` scripts should automatically format and print the `bioinfo.jobs.list` file to users' shell when they login.

For usage on the job submission script see [docs/utils/i-want-a-jobby.md]. The profile and zlogin scripts should simply work if placed in /etc/profile and /etc/zsh/zlogin, respectively.